% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblemap.R
\name{bubblemap}
\alias{bubblemap}
\title{Plot Points on a Map using D3.js}
\usage{
bubblemap(data, latcol = NULL, loncol = NULL, colorcol = NULL,
  namecol = NULL, sizecol = NULL, size = 5, mapdata = "usa",
  maxdomain = 1000, maxrange = 10, mapscale = 1280, legend = TRUE,
  top = 20, right = 40, bottom = 30, left = 50, width = NULL,
  height = NULL)
}
\arguments{
\item{message}{message}
}
\description{
D3.js provides the ability to generate beautiful maps using many projections.
}

