% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblemap.R
\name{bubblemapOutput}
\alias{bubblemapOutput}
\title{Widget output function for use in Shiny}
\usage{
bubblemapOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

