% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblemap.R
\name{renderBubblemap}
\alias{renderBubblemap}
\title{Widget render function for use in Shiny}
\usage{
renderBubblemap(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

